Add residents modaland implement closing and opening it
Modal only contains dummy data, for now.
It can be opened from the parentElement (the planets table component) via a prop. But closing is controlled from inside the modal (the child), so I also needed to pass back the closed state to the parent via a function (passed down as a prop).
This was necessary in order for the modal to be able to be opened from the parent again.